+++ core/sqldatabase/src/main/java/dk/groupa/sqldatabase/repository/ReserveRepository.java
@@ -13,11 +13,11 @@ import java.util.List;
 public interface ReserveRepository extends JpaRepository<BorrowQueue, Long> {
 
     @Query(value = "EXEC sp_ReserveBook @user_id = :user_id, @book_id = :book_id", nativeQuery = true)
-    BorrowQueue reserveBook(@Param("user_id") int user_id, @Param("book_id") int book_id);
+    BorrowQueue reserveBook(@Param("user_id") Long user_id, @Param("book_id") Long book_id);
 
     @Query(value = "SELECT * FROM borrow_queue", nativeQuery = true)
     List<BorrowQueue> getQueues();
 
     @Query(value = "EXEC sp_NextReserve @borrow_queue_id = :borrow_queue_id, @user_id = :user_id, @book_id = :book_id", nativeQuery = true)
-    Loan borrowReserve(@Param("borrow_queue_id") int borrow_queue_id, @Param("user_id") int user_id, @Param("book_id") int book_id);
+    Loan borrowReserve(@Param("borrow_queue_id") Long borrow_queue_id, @Param("user_id") Long user_id, @Param("book_id") Long book_id);
 }
